requirements.txt
absl-py==0.11.0
aiohttp==3.6.3
astor==0.8.1
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.3.0
blinker==1.4
brotlipy==0.7.0
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
contrib==0.3.0
cryptography==3.3.1
cycler==0.10.0
decorator==4.4.2
dlib==19.21.1
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.31.0
h5py==2.10.0
idna==2.10
idna-ssl==1.1.0
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==2.0.0
jsonschema==3.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.3.0
Markdown==3.3.3
matplotlib==3.3.2
multidict==4.7.6
networkx==2.5
numpy==1.15.3
oauthlib==3.1.0
olefile==0.46
opencv-contrib-python==4.4.0.46
opencv-python==4.4.0.46
opt-einsum==3.1.0
pandas==0.22.0
pid-controller==0.2.0
Pillow==8.0.1
pip==20.3.3
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pygame==2.0.0
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scipy==1.1.0
Shapely==1.7.1
six==1.15.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorboardX==2.1
termcolor==1.1.0
tifffile==2020.9.3
tornado==6.1
